Correcting a mistake
The previous code, fibonacci sequence would not go until 100, it would stop at 80, whereas the fibonacci sequence had to go on until 100. The reason is coz the sum should only contain the even numbers until 100. However, when I put until 100, then the fibonacci sequence goes until 144, so now in this I added an extra if condition to make this possible.